+++ examples/tauri-app/src/vite-env.d.ts
@@ -1,2 +1,7 @@
-/// <reference types="svelte" />
 /// <reference types="vite/client" />
+
+declare module "*.vue" {
+  import type { DefineComponent } from "vue";
+  const component: DefineComponent<object, object, unknown>;
+  export default component;
+}